ch24-functions

+++ virtual-machine/chunk.c
@@ -27,7 +27,7 @@ void writeChunk(Chunk* chunk, uint8_t byte, int line) {
 
 void freeChunk(Chunk* chunk) {
     FREE_ARRAY(uint8_t, chunk->code, chunk->capacity);
-    FREE_ARRAY(uint8_t, chunk->lines, chunk->capacity);
+    FREE_ARRAY(int, chunk->lines, chunk->capacity);
     freeValueArray(&chunk->constants);
     initChunk(chunk);
 }
